Add conversion tool from untarred format to pack
This is a conversion tool build as a separate executable that can
convert form the deduplicated format (what you get by untarring
the .tar.zst archives) to the pack format. The index that will be
created by this tool is suboptimal, as it is simply a list of time files
in the source directory, and as such it does not ripe the benefits
of the tree representation.

The metadata that is written is sufficient to decompress a specific
path, but is not what we want to have in the end. There would need
to be an additional way for us to specify which files make up individual
builds (commits, releases, snapshots, whichever one we choose to use).

This tool also allows for some self-validation by immediately
decompressing a path using the runtime representation.

Format code

Ran cargo fmt with default settings.

Apply clippy suggestions

Fix semver warning

Refactor code

Implemeted the suggested changes.

Signed-off-by: Peter Waller <peter.waller@arm.com>