Optimise path index and restructure code
- A new command line parameter was added `--strip-components`. It takes
a numeric value specifying how the separate builds are stored on disk.
For example: `--strip-components 2` will expect individual full builds
to be available at $INPUT/$dirA/$dirB/... The paths stored in the tree
will be relative to dirB.

- Compression level can be specified on the command line with
`--compression-level #`

- The contents of separate packed builds are now referred to as a
snapshot and each snapshot is identified by a tag. The tag is computed
by concatenating the named of the directories `--strip-components`
levels deep. Example: $dirA$dirB

- Each snapshot stores either a complete list of files, or a set of
changes relative to the previous one. The snapshots are stored in a
alphabetical order, based on their tag.

- The `--validate` parameter now expects a snapshot tag and unpacks the
contents of that snapshot under $OUTPUT.dir/

- Some new package dependencies were added: serde - serialisation lib;
rmp-serde - msgpack backend for serde; num_cpus - utility to get the
num of available CPUs at runtime (already a dependency of rayon).
These are all available under MIT license.

Signed-off-by: Peter Waller <peter.waller@arm.com>