Add elfshaker store and elfshaker list
This commit adds two new subcommands to the existing elfshaker
executable.

The elfshaker list command can be used to list the contents of a repo,
a pack or a snapshot, depending on the options used.

The elfshaker store command can be used to capture a snapshot of the
current state of the repository. The snapshots captured in this way are
stored in the elfshaker_data/unpacked directory and not in a pack file.
The ability to create a pack file from the snapshots captured in this
way will have to be introduced separately.

Snapshot tags and pack names must now match the regex pattern
[a-zA-Z_\-]. This was done to make parsing easier, and to allow for
special syntax to be introduced in the future. This pattern can be made
less restrictive in the future. A special pack name, "unpacked", is now
also reserved and refers to the group of unpacked snapshots captured by
elfshaker store.

Some of the existing file structure was changed, particularly the
repo module has been broken up into several separate modules, to
promote code reuse and extensibility.

References to snapshots and packs are now represented by custom types,
which lead to further changes in the existing code.

Signed-off-by: Peter Waller <peter.waller@arm.com>