Fix incorrect entry diffing
This fixes the failure reported when extracting a snapshot when --reset
is not specified.

The failure was caused by an incorrect file entry diffing
implementation, which did not take into account the case where
multiple copies of the same file could be present, and consequently
resulted in the deletion of the wrong file from the filesystem.

This caused successive extract operations to fail with DirtyWorkDir.

Signed-off-by: Peter Waller <peter.waller@arm.com>