Add new custom compression format support
Added support for a new .pack format backend that splits objects
into separate frames. This leads to performance increases when the
objects that need to be extracted a co-located in the same frame,
since only that frame needs to be decompressed, thereby saving on
seeking.

The .pack file format is a superset of the Zstandard format and
contains an elfshaker header in a skippable frame at the beginning of
the file, followed by N compressed frames.

elfshaker pack and create-pack now support a `--frames` option to
specify the number of frames to emit in the pack (default is input size
/ 512 MiB).

The format is of the .pack file (newer/older format) is auto-detected.

Signed-off-by: Peter Waller <peter.waller@arm.com>