Use timestamps to detect changes, not permissions
This commit changes the heuristic used for detecting changes in the
files extracted by elfshaker to use timestamps (latest of mtime, btime
on Unix platforms). The timestamps are compared to the timestamp of
the current HEAD. If any of the files extracted from the HEAD snapshot
is more recent than HEAD, extract will fail. This new heuristic acts
as a replacement the previous one, which necessitated extracted files
to be extracted with read-only permissions and verified that those are
unchanged before overwriting the files with the contents of a new
snapshot.

With this commit, files no longer have their write permission stripped
after extraction. The checks introduced by this commit can be disabled
with `--force` as before.

Signed-off-by: Peter Waller <peter.waller@arm.com>