Rework PathPool
This is a reimplementation of PathPool to use a flat array of strings to
represent paths, rather than the path tree. This results in a slightly
less efficient representation of paths but simplifies things quite a
bit. We could regain some of the efficiency by packing the index later.

Fix #20.

This changes the pack format, so update verification.pack.

Signed-off-by: Peter Waller <peter.waller@arm.com>