Rework check.sh
I discovered that check was failing silently in a couple of cases so
took a closer look.

I've sanity checked it a bit further, including writing incorrect data
out and checking that it is picked up, so I have greater confidence the
test harness is working as intended.

* `set -euo` flags etc, weren't passed to subshells. `export SHELLOPTS` fixes that.
* diff the full list of files on the filesystem, rather than just
  checking files reported by elfshaker, so that extraneous files are
  detected.
* the syntax `[ ! "elfshaker" ... ]` was incorrect, and was returning a
  non-zero exit status because `[` itself was failing, not because the
  condition was 'good'. This hid the fact that the program was not
  exiting with an error code.
* use /dev/shm tmpfs (my $WORKDIR is NFS which is a little slow)
* use --compression-level=1 for speed.
* remove sleeps (did they have a purpose? If so please document.)
* show full context on failure, not just last 25 lines

The next patch will fix some issues with error handling that check.sh
now picks up.

Signed-off-by: Peter Waller <peter.waller@arm.com>