Make check.sh respect -e in tests
It turns out that `set -e` is ignored, even in functions and subshells if the
invocation appears within an `if`.

For more information search the web for 'surprising errexit behaviour conditional'.

Flags are propagated into the subshells, though, so my previous `export
SHELLOPTS` fix was incorrect (and I had not identified the root cause of the
problem).