Make {PathPool => EntryPool} generic
I've noticed that we're effectively using the same trick for Objects as we were
using for Paths, that is, we intern them keyed on their checksum, so that we
can refer to them via a shorter identifier. So let's make PathPool generic, and
use it for objects.

This is a stepping stone, renaming PathPool to EntryPool and updating existing
uses of the former. Updating Objects will come later and is more invasive. At
the moment we won't end up with distinct types for Handle, which is
unfortunate. I'm thinking that this can be fixed later, possibly by introducing
another type parameter to the EntryPool generic identifying distinct Handle
types.

Signed-off-by: Peter Waller <peter.waller@arm.com>