Use EntryPool to store objects
This reworks the storage mechanism so that appending objects to a pack works in
a similar way to appending paths. Objects no longer need renumbering in order
to append to a pack.

They no longer have an intrinsic ordering until we decide an order for them,
which we store separately.

Writing the object offsets in compute_object_offsets_and_ordering takes a 'mut
self' parameter to indicate that this is a mutating operation.

Signed-off-by: Peter Waller <peter.waller@arm.com>