Delete empty directories (#69)
Implemented the git way of handling directories which are no longer
(implicitly) referenced: Attempt to remove the directory, since we no
longer need it, but if that fails (because the directory contains
files), we leave the directory alone.

The above is done by optimistically calling `rmdir()`, ignoring
the potential `ENOTEMPTY` (or the platform equivalent), and repeating
this recursively until the repository root is reached.

We reduce the number of system calls by using an abstract queue which
employs a heuristic dependent on the directory list being sorted.
A sorted directory list means that parent directories come before child
directories. Enqueuing a directory which is a child of the last enqueued
directory can be reduced to only storing the child. Enqueuing a non-child
eagerly calls the removal procedure on the last enqueued directory first.

Signed-off-by: Veselin Karaganev <veselin.karaganev@arm.com>