Fix atomicfile race condition
The atomicfile test was occasionally failing because it was possible that a
process could come along and delete the target file 'as stale' in the gap
between the file being created and the lock being taken. This would leave a
process on the happy file with a locked file descriptor which was unlinked.

The fix is to improve the locking: when the lock is taken, check it still
corresponds to the linked file on disk. If it doesn't, then the race was lost,
and this is treated as a failed locking operation.

With this fix I haven't yet seen the tests fail on Linux or OSX.

Fixes #47.