Introduce load_index_snapshots
This reduces the CPU time for list, find, extract from >3s to 10ms in my test
repository by only loading the snapshot list from the beginning of the index.

`load_index_snapshots` returns a Vec<String> rather than a partial `PackIndex`
so that it can't be accidentally used where a full `PackIndex` is expected.

Updates #9.