Add `elfshaker clone`
`elfshaker clone` clones a remote repository into a new directory.
It does that by fetching a remote `.esi` (over HTTP) and creating a
remote called origin. After fetching the `.esi`, the command proceeds
to fetch all of the available `.pack.idx` from the remote, so that
`elfshaker update` does not have to be run manually.

Signed-off-by: Veselin Karaganev <vesko.karaganev@gmail.com>