Workaround git log --since missing commits
Git's walking mechanism stops as soon as it sees a commit which is older
than the date specified by --since.

This is problematic in LLVM's commit history in August 2022, because
there is a commit in the middle of it with a CommitDate in 2021.

Work around this by considering a large number of commits and taking the
oldest one matching the target date.

Fix #89.

Signed-off-by: Peter Waller <peter.waller@arm.com>