Refactor: Version the PackIndex structure
Separated the PackIndex definition from its implementation into a
PackIndex trait and PackIndexV1 struct.

Also added a `VerPackIndex` (versioned pack index) enum which will allow
us to evolve the pack index format in the future, by handling the differences
between the versions of the `PackIndex` in the `PackIndex` impl of the
`VerPackIndex` enum.