Relax the repository lock
Writing to disk using elfshaker store is safe to run concurrently
becase we move the object into the loose object storage using atomic
FS operations (rename(2)). We stored the objects in a content
addressable way, and we check if the object exists before creating it.

Because that check-then-copy operation is not atomic, it is not safe to
delete objects while another elfshaker process is writing objects into
the object storage.

Repositry objects now take a shared lock on the repository directory,
(blocking if an exclusive lock is held). The shared lock is upgraded
to an exclusive lock the first time we try to perform a destructive
operation such as deleting a pack or an object.

The file lock is _not_ downgraded to a shared lock automatically and
is unlocked when the Repository object is dropped. For the purposes
of the elfshaker CLI, this coincides with when the process exists.