Use hashbrown::HashMap for the ChangeSet
Consistent runtime improvement when running elfshaker gc from
26s -> 18s (on my machine and test repository based on manyclangs) due
to the overhead of applying `ChangeSet<FileHandle>` for ~60k snapshots.

The hashbrown implementation is a collections::HashMap replacement with
SIMD lookup and quadrating probing. It uses a weaker but faster hash
function and is both lower in memory usage and much faster than the
collections::HashMap.