[manyclangs] Include llvm-min-tblgen in the set of binaries written
To reduce the disk IO writes - which can be a bottleneck when building a
pack - most binary outputs are rewritten to send their outputs to
/dev/null.

However, we need some binaries (llvm-tblgen for example).

Since https://reviews.llvm.org/D146352 (beginning of the May pack, which
I haven't yet built), there is another binary called llvm-min-tblgen
which is also required. Include it here.