Ignore ./elfshaker_data while finding files
We had an existing mechanism to do this, but it does not work well on
windows+wine, because the paths become rather complicated.

Early during walk, reject */elfshaker_data. We should be the only users
of this name so it should have minimal impact.

Ideally this logic would just reject ./elfshaker_data and it should be
updated to that but I'm fine with this workaround for now while getting
windows test infra up and running.