Add macos-14 (aarch64) darwin builder
Should give aarch64 darwin binaries.

Github's builder naming is confusing. They had -arm64 suffices, but now
they don't, so it's unclear how to use their labels to specify which
architecture you want to target. For now we'll assume macos-13 = x86 and
macos-14 = aarch64 and in the future we should be able to do cross
compilation to x86 from aarch64 if it works easily enough (dependent on
nixpkgs).