contrib/elfshaker-serve: Init
Add an elfshaker-serve daemon which serves binaries tarballs out of elfshaker snapshots.

* It can cache the linked binaries to S3 and serve a presigned URL to
  the client to directly fetch from S3.
* It rate limits to prevent total CPU starvation.
* It is able to map any length prefix of a commit onto a snapshot.
* It's configurable through command line options.

The basic syntax is curl localhost:8080/c/<commit>/<binary>.

The server must be provided with a plaintext list of snapshots, which it
identifies through the final part of snapshot name (split on -),
assuming this is a commit hash.

It calls out to a shell script called extract-and-link.sh, which is run
from a temporary directory. This shell script must be supplied
seperately. Its job is to run elfshaker extract and link the binary.

Once the script has run it generates a tarball of the linked binary.

It has special case logic in it for clang to package up the resource
headers.

It has additional logic to see through symlinks (so if you request
llvm-readelf which is a symlink to llvm-readobj, you get both in the bin
dir of the resulting tarball).